requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
pyyaml
psutil